+++ server/package.json
@@ -37,6 +37,7 @@
     "@types/cors": "^2.8.13",
     "@types/csurf": "^1.11.2",
     "@types/express": "^4.17.17",
+    "@types/jsonwebtoken": "^9.0.2",
     "@types/lodash.camelcase": "^4.3.7",
     "@types/lodash.snakecase": "^4.1.7",
     "@types/morgan": "^1.9.4",
