--- server/utils/auth.js
@@ -1,62 +1,0 @@
-require("dotenv").config();
-const jwt = require("jsonwebtoken");
-const jwtSecret = process.env.JWT_SECRET;
-const jwtExpiresIn = process.env.JWT_EXPIRES_IN;
-const User = require("../models/User");
-
-const setTokenCookie = (res, user) => {
-  // Create the token.
-  const token = jwt.sign(
-    { data: user },
-    jwtSecret,
-    { expiresIn: parseInt(jwtExpiresIn) } // 604,800 seconds = 1 week
-  );
-
-  const isProduction = process.env.NODE_ENV === "production";
-
-  // Set the token cookie
-  res.cookie("token", token, {
-    maxAge: jwtExpiresIn * 1000, // maxAge in milliseconds
-    httpOnly: true,
-    secure: isProduction,
-    sameSite: isProduction && "Lax",
-  });
-
-  return token;
-};
-
-const restoreUser = (req, res, next) => {
-  const { token } = req.cookies;
-  req.user = null;
-
-  return jwt.verify(token, jwtSecret, null, async (err, jwtPayload) => {
-    if (err) {
-      return next();
-    }
-
-    try {
-      const { id } = jwtPayload.data;
-      req.user = await User.getSession(id);
-    } catch (e) {
-      res.clearCookie("token");
-      return next();
-    }
-
-    if (!req.user) {
-      res.clearCookie("token");
-    }
-
-    return next();
-  });
-};
-
-// If there is no current user, return an error
-const requireAuth = (req, _res, next) => {
-  if (req.user) return next();
-
-  const err = new Error("Authentication required");
-  err.status = 401;
-  return next(err);
-};
-
-module.exports = { setTokenCookie, restoreUser, requireAuth };
